corrected bug where we could only have one @@ and one prog (now either both progs have @@ or none have), next commit will do with general improvements and small bug fixes

+++ Instrumentation/afl-instr/afl-fuzz.c
@@ -91,7 +91,6 @@ short test_bool = 0;
 
 EXP_ST u8 *in_dir,                    /* Input directory with test cases  */
           *out_file,                  /* File to fuzz, if any             */
-		  *out_file_delta[MAX_AMOUNT_OF_PROGS],
           *out_dir,                   /* Working & output directory       */
           *out_dir_delta[MAX_AMOUNT_OF_PROGS], /* Each program should have their own out dir, this so far is only for testing*/
           *sync_dir,                  /* Synchronization directory        */
@@ -2297,13 +2296,14 @@ EXP_ST void init_forkserver_special(char** argv, u8 **path, s32 *forksrv_pid,
     	close(out_fd);
 
     }
+    
     /*
     for(int i = 0; i < numbr_of_progs_under_test; i++){
-    	if (out_file_delta[i]) {
-    		dup2(dev_null_fd, 0);
+    	if ( out_file_delta[i]) {
+    		//dup2(dev_null_fd, 0) ;
     	} else {
       		//enters here
-    		dup2(out_fd_delta[i], 0);
+    		//dup2(out_fd_delta[i], 0);
     		close(out_fd_delta[i]);
     	}
     }
@@ -2871,6 +2871,7 @@ static void write_to_testcase(void* mem, u32 len) {
 
 		} else close(fd);
 		
+		// used mostly to when we have  @@ in args
 		/*
 		s32 fd = out_fd_delta[CUR_PROG];
 
@@ -2902,7 +2903,7 @@ static void write_with_gap(void* mem, u32 len, u32 skip_at, u32 skip_len) {
   s32 fd = out_fd;
   //s32 fd = out_fd_delta[CUR_PROG];
   u32 tail_len = len - skip_at - skip_len;
-
+  
   if (out_file) {
 
     unlink(out_file);
@@ -5019,8 +5020,8 @@ static void show_stats(void) {
        DTD(cur_ms, start_time), tmp, DI(queue_cycle - 1));
 
    SAYF(bV bSTOP "    cur run time : " cRST "%-34s " bSTG bV bSTOP
-       "  cur prog : %s  " bSTG bV "\n",
-       DTD(cur_ms, prog_start_time), prog_names[CUR_PROG] );
+       "  cur prog : %d  " bSTG bV "\n",
+       DTD(cur_ms, prog_start_time), CUR_PROG );
 
   //SAYF(bV bSTOP "   current program : " cRST  "%-34s\n" bSTG bV bSTOP, cur_prog_title);
 
@@ -8034,14 +8035,13 @@ static void getProgsBlockList(){
 		//fblocks = fopen("./progs_blocks.txt","r");
 	  	fblocks = fopen(path_instr,"r");
 
-	  	free(path_instr);
 
 		if( fblocks == NULL ){
 			// no list of blocks
 			FATAL("No blocks list found with name %s", path_instr);
 		}
 		
-
+		free(path_instr);
 		while ( getline(&line, &len, fblocks) != -1 ){
 			//printf("beggin\n");
 	    	unsigned short block_id = atoi(line);
@@ -8494,31 +8494,34 @@ EXP_ST void setup_all_dirs_fds(void) {
 EXP_ST void setup_stdio_file(void) {
 
 	//printf("\t-> in setup_stdio_file out_dir -> %s\n", out_dir);
+	if( !out_file ){
+		u8* fn = alloc_printf("%s/.cur_input", out_dir);
 
-	u8* fn = alloc_printf("%s/.cur_input", out_dir);
+	  unlink(fn); /* Ignore errors */
 
-  unlink(fn); /* Ignore errors */
+		out_fd = open(fn, O_RDWR | O_CREAT | O_EXCL, 0600);
 
-	out_fd = open(fn, O_RDWR | O_CREAT | O_EXCL, 0600);
+		if (out_fd < 0) PFATAL("Unable to create '%s'", fn);
 
-	if (out_fd < 0) PFATAL("Unable to create '%s'", fn);
+		ck_free(fn);
+	}
+	/*
+    for(int i = 0; i < numbr_of_progs_under_test; i++){
 
-	ck_free(fn);
+    	//printf("\t-> in setup_stdio_file out_dir -> %s\n", out_dir_delta[i]);
+    	if( out_file_delta[i] ) continue;
 
-  for(int i = 0; i < numbr_of_progs_under_test; i++){
-    //printf("\t-> in setup_stdio_file out_dir -> %s\n", out_dir_delta[i]);
+    	u8* fn = alloc_printf("%s/.cur_input", out_dir_delta[i]);
 
-    fn = alloc_printf("%s/.cur_input", out_dir_delta[i]);
+    	unlink(fn); 
 
-    unlink(fn); /* Ignore errors */
+    	out_fd_delta[i] = open(fn, O_RDWR | O_CREAT | O_EXCL, 0600);
 
-    out_fd_delta[i] = open(fn, O_RDWR | O_CREAT | O_EXCL, 0600);
+    	if (out_fd_delta[i] < 0) PFATAL("Unable to create '%s'", fn);
 
-    if (out_fd_delta[i] < 0) PFATAL("Unable to create '%s'", fn);
-
-    ck_free(fn);
+    	ck_free(fn);
   }
-
+	*/
 }
 
 
@@ -8831,7 +8834,7 @@ EXP_ST void detect_file_args(char** argv, int prog) {
       if (!out_file)
         out_file = alloc_printf("%s/.cur_input", out_dir);
 
-      //if (!out_file_delta[prog])
+      //if ( !out_file_delta[prog] )
       //  out_file_delta[prog] = alloc_printf("%s/.cur_input", out_dir_delta[prog]);
 
       /* Be sure that we're always using fully-qualified paths. */
@@ -8850,9 +8853,61 @@ EXP_ST void detect_file_args(char** argv, int prog) {
       *aa_loc = '@';
 
       if (out_file[0] != '/') ck_free(aa_subst);
-      //if (out_file_delta[prog][0] != '/') ck_free(aa_subst);
+      //if ( out_file_delta[prog][0] != '/' && aa_subst ) ck_free(aa_subst);
 
-    }
+    } // end of if found @@
+
+    i++;
+
+  }
+
+  free(cwd2); /* not tracked */
+
+}
+
+/* Detect @@ in args. */
+/* Marks where the file must go*/
+/*TODO -> need to make sure that if one has @@ all of them have it*/
+EXP_ST void detect_file_args_delta(char** argv, int prog) {
+
+  u32 i = 0;
+  u8* cwd2 = getcwd(NULL, 0);
+
+  u8 *change;
+
+  if (!cwd2) PFATAL("getcwd() failed");
+
+  while (argv[i]) {
+
+    u8* aa_loc = strstr(argv[i], "@@");
+
+    if (aa_loc) {
+
+      printf("prog %d has @@\n", prog);
+
+      u8 *aa_subst, *n_arg;
+
+      /* If we don't have a file name chosen yet, use a safe default. */
+
+      
+      change = alloc_printf("%s/.cur_input", out_dir_delta[prog]);
+
+      /* Be sure that we're always using fully-qualified paths. */
+
+      if (change[0] == '/') aa_subst = change;
+      else aa_subst = alloc_printf("%s/%s", cwd2, change );
+
+      /* Construct a replacement argv value. */
+
+      *aa_loc = 0;
+      n_arg = alloc_printf("%s%s%s", argv[i], aa_subst, aa_loc + 2);
+      argv[i] = n_arg;
+      *aa_loc = '@';
+
+      //if (out_file[0] != '/') ck_free(aa_subst);
+      if ( change[0] != '/' && aa_subst ) ck_free(aa_subst);
+
+    } // end of if found @@
 
     i++;
 
@@ -9266,10 +9321,13 @@ int main(int argc, char** argv) {
   if (!timeout_given) find_timeout();
 
   detect_file_args(argv, 0);
-  for(int i = 0; i < numbr_of_progs_under_test; i++)
-  	detect_file_args(prog_args[i], i);	
+  for(int i = 0; i < numbr_of_progs_under_test; i++){
+  	detect_file_args_delta(prog_args[i], i);	
+  }
 
-  if (!out_file) setup_stdio_file();
+  //if (!out_file) setup_stdio_file();
+  setup_stdio_file();
+
   //printf("%s\n", tmp_test);
 
   start_time = get_cur_time();
@@ -9280,7 +9338,7 @@ int main(int argc, char** argv) {
 
   /*we start forkservers here*/
   init_all_forkservers(argv);
-  get_prog_targets(0,1); //TODO - not working
+  get_prog_targets(0,1); 
 
   if (qemu_mode)
     use_argv = get_qemu_argv(argv[0], argv + optind, argc - optind);
@@ -9302,6 +9360,13 @@ int main(int argc, char** argv) {
 
   if (stop_soon) goto stop_fuzzing;
 
+  for(int pgr = 0; pgr < numbr_of_progs_under_test; pgr++){
+    printf("args %d = ", pgr);
+    for(int i = 0; prog_args[pgr][i] ; i++){
+    	printf("%s ", prog_args[pgr][i]);
+    }
+    printf("\n");
+  }
 
   /* Woop woop woop */
